made version parsing work better
Added parse_name_of_bzball method to SRP's utils module.  This is now used
to determine the version string for a built brp instead of just looking for
the first tak-separated segment that starts with a digit.

The new method works by parsing through the source tarball's filename in
reverse order after splitting on '-'.  The index of the latest "version"
segment is tracked until the first "non-version" segment.  For this method,
is_version test just looks for . in the string.  We might need to elaborate
on this later...