srp: add SRP_NOINSTALL_LA flag
This commit adds the SRP_NOINSTALL_LA flag.  This flag tells srp to ignore
any libtool libraries that get staged to be installed.  Files are determined
to be libtool libraries by parsing the output of the 'file' command.

SRP_NOINSTALL_LA was added as a default flag, so a SRP_NO_NOINSTALL_LA flag
was also added.