fixed dependency mapping loop
This loop was iterating over every line in the log, expecting it to be a
filename...  which was only true eons ago.  Tons of SRP_FLAGS add extra
lines to the log, which this loop has to skip over.  This problem didn't
really show up until I did some --verbose package building while hunting
down a different bug.